+++ vite.config.js
@@ -32,6 +32,9 @@ export default defineConfig( {
                 ],
             },
             workbox: {
+                skipWaiting: true,
+                clientsClaim: true,
+                cleanupOutdatedCaches: true,
                 navigateFallback: `/index.html`,
                 globIgnores: [ `**/*.wasm` ],
                 runtimeCaching: [